requirements.txt
torch==2.0.1 
torchvision==0.15.2 
numpy>=1.22.4 
yacs>=0.1.8 
opencv-python>=4.6.0 
scikit-learn>=1.0.2 
matplotlib>=3.5.2 
pillow>=9.2.0 
timm>=0.6.12 
tensorboard>=2.9.1 
tqdm>=4.64.0 
pandas>=1.4.3 
pyyaml>=6.0 
scipy>=1.8.1 
kafka-python>=2.0.2 
pyspark>=3.3.0 
faiss-gpu>=1.7.2 

